python/requirements.txt
# Face Recognition Service Dependencies
# Note: dlib is installed separately in the install.bat script
opencv-python>=4.8.0,<4.9.0
face-recognition>=1.3.0,<1.4.0
flask>=2.3.0,<2.4.0
flask-cors>=4.0.0,<4.1.0
pillow>=10.0.0,<10.1.0
numpy>=1.24.0,<1.25.0
